+++ packages/embed/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @opencx/widget
 
+## 4.0.31
+
+### Patch Changes
+
+- Rename `specialComponents` to `customComponents`
+
 ## 4.0.30
 
 ### Patch Changes